Remove Storage trait
Now that we only have one implementation it doesn't make sense to keep
the trait. If we end up wanting to mock it we can still add it back.
